refactor: AvatarGroupAnimation
