+++ components/dumb/Avatar/GroupAnimation/Roll.js
@@ -1,8 +1,8 @@
 import AvatarGroup from '@mui/material/AvatarGroup';
 import styled from '@mui/material/styles/styled';
 import { keyframes } from '@emotion/react';
-import { useCallback, useState } from 'react';
-import Button from '@mui/material/Button';
+import { animationPaused, prefersReducedMotionNoAnimation } from 'helpers/styles/animation';
+import AvatarGroupAnimation from 'components/dumb/Avatar/GroupAnimation';
 
 const roll = keyframes`
   0% {
@@ -34,28 +34,14 @@ const AvatarGroupAnimated = styled(AvatarGroup)({
     '&:first-of-type': {
       animation: `${roll} 3s ease infinite 2s`,
     },
-    '@media (prefers-reduce-motion)': {
-      animation: 'none !important',
-    },
+    ...prefersReducedMotionNoAnimation(),
   },
-}, (props) => props.paused && ({
-  '& .MuiAvatar-root': {
-    animationPlayState: 'paused !important',
-  },
+}, (props) => ({
+  '& .MuiAvatar-root': animationPaused(props),
 }));
 
-const AvatarGroupAnimationRoll = (props) => {
-  const [paused, setPaused] = useState(false);
-
-  const onToggle = useCallback((e) => {
-    e.stopPropagation();
-    setPaused((prev) => !prev);
-  }, [setPaused]);
-  return (
-    <Button variant="text" color="inherit" onClick={onToggle}>
-      <AvatarGroupAnimated paused={paused} {...props} />
-    </Button>
-  );
-};
+const AvatarGroupAnimationRoll = (props) => (
+  <AvatarGroupAnimation component={AvatarGroupAnimated} {...props} />
+);
 
 export default AvatarGroupAnimationRoll;
